+++ jax/_src/config.py
@@ -286,6 +286,47 @@ class Config:
 
     return _StateContextManager(name, help, update_thread_local_hook, validate)
 
+  def define_int_state(
+      self, name: str, default: Optional[int],
+      help: str, update_global_hook: Optional[Callable[[str], None]] = None,
+      update_thread_local_hook: Optional[Callable[[Optional[str]], None]] \
+        = None):
+    """Set up thread-local state and return a contextmanager for managing it.
+    Args:
+      name: string, converted to lowercase to define the name of the config
+        option (and absl flag). It is converted to uppercase to define the
+        corresponding shell environment variable.
+      enum_values: list of strings representing the possible values for the
+        option.
+      default: optional int, default value.
+      help: string, used to populate the flag help information as well as the
+        docstring of the returned context manager.
+    Returns:
+      A contextmanager to control the thread-local state value.
+    See docstring for ``define_bool_state``.
+    """
+    name = name.lower()
+    default_env = os.getenv(name.upper(), default)
+    if default_env is not None:
+      try:
+        default = int(default_env)
+      except ValueError:
+        raise ValueError(f"Invalid value \"{default_env}\" for JAX flag {name}")
+    self.DEFINE_integer(name, default, help=help, update_hook=update_global_hook)
+    self._contextmanager_flags.add(name)
+
+    def get_state(self):
+      val = getattr(_thread_local_state, name, unset)
+      return val if val is not unset else self._read(name)
+    setattr(Config, name, property(get_state))
+
+    def validate(new_val):
+      if new_val is not None and not isinstance(new_val, int):
+        raise ValueError(f'new int config value must be None or of type int, '
+                         f'got {new_val} of type {type(new_val)}')
+
+    return _StateContextManager(name, help, update_thread_local_hook, validate)
+
   def define_string_state(
       self, name: str, default: Optional[str], help: str,
       update_global_hook: Optional[Callable[[str], None]] = None,
@@ -709,6 +750,16 @@ raise_persistent_cache_errors = config.define_bool_state(
           'caught and raised as warnings, allowing program execution to '
           'continue. Defaults to false so cache bugs or intermittent issues '
           'are non-fatal.'))
+
+persistent_cache_min_instruction_count = config.define_int_state(
+    name='jax_persistent_cache_min_instruction_count',
+    default=6,
+    help=('The minimum number of instructions a computation needs to have to '
+          'be written to the persistent compilation cache. This threshold can '
+          'be raised to decrease the number of entries written to the cache. '
+          'The (unoptimized) instruction count is meant to be a proxy for '
+          'compile time, so programs with longer compile times are still '
+          'cached.'))
 
 hlo_source_file_canonicalization_regex = config.define_string_state(
     name='jax_hlo_source_file_canonicalization_regex',